Bondage club room previews (#84)
* More SlaveTwin

* To cut or not to cut

* SlaveTwin Beta Ready

The SlaveTwin section should be ready for some testing, but the accompanying RopeGroup section still needs a lot of work doing to it.

* Cleanup random selection function

This still uses the previous random number selection method, it's just been cleaned up to make it more readable and to avoid duplication.

As far as I could tell, there was no reason for the `C101_KinbakuClub_Slaves_Random#` variables to be global. So that's been incorperated into the function to keep things cleaner.

* RopeGroup tidied for Alhpa

* Change going to K Club.

* More KC fitting in

* Can leave KC

Changed the door in the club, player can now leave if not in bondage.

* A bit more RopeGroup

* Twins Stage 1

First part of main interaction with either twin.

* upload 14-11-2018

* Update Cassi 1

* Return to dorm added

* RG bug fix

Fixed reference error

* Cassi with rope group release

* Disabled save function in club

* Nursery and Market club room introductions

Previews of the nursery and slave market for the bondage club.

* This modification is probably a mistake

Post is again if it wasn't a mistake, I'm not sure why it's there.