Friend list in chat search (#265)
* More SlaveTwin

* To cut or not to cut

* SlaveTwin Beta Ready

The SlaveTwin section should be ready for some testing, but the accompanying RopeGroup section still needs a lot of work doing to it.

* Cleanup random selection function

This still uses the previous random number selection method, it's just been cleaned up to make it more readable and to avoid duplication.

As far as I could tell, there was no reason for the `C101_KinbakuClub_Slaves_Random#` variables to be global. So that's been incorperated into the function to keep things cleaner.

* RopeGroup tidied for Alhpa

* Change going to K Club.

* More KC fitting in

* Can leave KC

Changed the door in the club, player can now leave if not in bondage.

* A bit more RopeGroup

* Twins Stage 1

First part of main interaction with either twin.

* upload 14-11-2018

* Update Cassi 1

* Return to dorm added

* RG bug fix

Fixed reference error

* Cassi with rope group release

* Disabled save function in club

* Nursery and Market club room introductions

Previews of the nursery and slave market for the bondage club.

* This modification is probably a mistake

Post is again if it wasn't a mistake, I'm not sure why it's there.

* Partial nursery update

* upload from desktop for fork update

* upload from desktop for fork update

* upload from desktop for fork update

* Most of Nurse interactins done

* Nursery part 1 ready

* Cloth Fix and First Lingerie

* Lingerie

* mittens partical success

* PR 19 merge changes and more mittens

* Nursery v1 Ready for Beta

First proper section of the nursery updated ready for testing

* InventoryGet(...) is null fix

* Vibrating nipple clamps (not finished) + strapless bra

* More Vibe clamps, fishnets and strait leotard

* own diaper fix

* Debuff + Engwishh

* fix + items

* Update Female3DCG.csv

* Nursery Cost fix + pigtial ribbon + bunnygirl set

* Pantyhose Priority

quick fix to layer priorities

* Carrot gag fix

* BunnyEar2 image fix

* Pony Ears

* Robe and captains hat

* Slave market all refrence to $ removed

* Update Skill.js

* Puppy and collar items

* Missing locks

* Bug fixes

* Access friendlist from chatsearch

Adds option to also view your friends list in the ChatSearch screen and return to ChartSearch only if you came from there instead of the information sheet.

I did wonder about extending the function CommonSetScreen() so that is could log the screen you came from and return you there next time, if a destination screen was not specified.  It might be a technically better solution than what I have added, but I'm not sure how worth while that might be unless there are more screens that become are accesible from multiple other screens.