Fixes for the Strait Dress and Removal of redundent ItemHides. (#1052)
* Update Female3DCG.js

Removed the HideItems from the MermaidSuit since all of them were ItemFeet which is already hidden via the normal Hide.

Added "LegsClosed" to any item that has "StraitDressOpen". Since there doesn't seem to be a reason not to and it will better respect functions looking for "LegsClosed" which "StraitDressOpen" is very similar to except that . If there is a reason not to then feel free to revert these the effect items are as follow. StraitDressOpen, SeamlessStraitDressOpen, Bitchsuit & BitchSuitExposed.

Removed "ItemFeetMermaidRopeTie" since the item no longer exists in game due to being merged with HempRope thus the hide is no longer needed.

* Update Female3DCG.js

Added the Ornate Cuffs to the Strait Dresses to prevent a conflict

* Update Female3DCG.js

Changed the OrnateAnkleCuffs to ItemFeetOrnateAnkleCuffs