Fixed Baby Harness
Replaced all images of the baby harness in different poses with the one from the base pose since it is greyscaled. This means that it is no longer different colours when binding people. Also added the harness to the OverHead pose that seems to have been forgotten, since it causes no issues.