Futuristic Collar (#1589)
* Futuristic item upgrades

First commit. Roadmap:
-Chastity bra will be a diagnostic item (mostly fluff)
-Chastity belt will have several functions to aid in training
-
-Futuristic items in general will give an error sound effect when a user who cannot unlock the item attempts to access the controls

Changes included in this commit:
-Bra implementation. Now displays heart rate and also a little heart icon when the user is orgasming or when the user is being aroused by an item action
-CommonDraw updated to pass G for size-based modifications
-Dialog added
-Assets added to futuristic bra to allow recoloring the text

* Added access denied effect to the panel gag

* Access denied sound effect

* Added ACCESS DENIED to the other futuristic items

Except the bra, which does not have any actual functionality besides appearance (but it has a fancy animation!)

* AlwaysExtend for all futuristic items as they are bluetooth enabled ^^

* Added ability to close the futuristic bra display + added muzzle gag

These changes should help players that prefer the metal look without too many displays

* Made the visor harder to take off when locked.

Also removed a debug statement in CommonDraw.js

* Added chat room message for trying to access a locked futuristic item

* Access Denied message now remains until the user exits

The load function kept removing the access denied message so now it's just called in the exit function

* Added a null check to prevent crash

* Futuristic Collar

Added a futuristic collar, which has the functionality of allowing a player to quickly lock all items with "Futuristic" or "Interactive" in their name on a player. It only allows players to use the following locks, and only allows the player to lock or unlock the items if they own the locks and could normally lock the player:
 - Metal Padlock
 - Exclusive Padlock
 - 5 Min Padlock
 - Mistress Padlock
 - Lovers Padlock
 - Owner Padlock


This has a dependency on the futuristic item revamp

* Fixed partial locking, other bugs

* Start

* Revert "Start"

This reverts commit 20f4e31f9597f8bbb71606520c210d5549e101d4.

* Added a hydraulic locking sound instead of the default lock/unlock

Recorded it myself so no attribution needed

* Included a new sound for putting the collar on

Also fixed sound effects not playing for others

* Fixed a bug that allowed you to put blocked locks on someone

Co-authored-by: Ben987 <30805159+Ben987@users.noreply.github.com>