Maid Button (#1564)
* Maid Button

This PR adds a button that the user must press before maids will arrive. It can be disabled with a toggle in the preference menu. The button simply starts the timer for the maids that normally starts if the user is restrained, and does not add additional functionality to the maids.

Button icon by Nina

* Typo

* Changed the button toggle to AutoMaid

* Fixed

* Made the (Nod Yes) and asking for help dialogue options not appear

* Changed AutoMaid to DisableAutoMaid

Also did some code improvements, added null checking, break, and MouseIn statements

* Made option a positive choice 

Checkbox enables the maids, rather than the other way. Default behavior is to enable maids, still.

* Added bell ring sound when ringing maid bell

* Made ring effect respect user volume