Password Lock (#1663)
* Password Lock

Adds a new lock, similar to the combination lock. However, instead of being limited to only numbers, it is limited to only 8 letter passwords made of letters, and the user can set a hint. Once a password is set, it cannot be changed.

Additionally, futuristic items will use the password entered by players and can be unlocked if the password entered in the field is the same.

* Removed MasterLock brand name

* Added validation to make sure lock properties are properly deleted

* When the item is unlocked it also deletes the password

* Fixed CombinationNumber instead of Password

* Fixed other wrongful occurrences of CombinationNumber

* Futuristic items now work with combination lock too