Increases mas password length to 8 (#1761)
This change was requested by players. I cant remember the reason it was set to 6, but I think it had to do with the possibility of trolls setting one to random letters and being unable to figure it out. However, bruteforcing 6 letters is still hard, and in Extreme mode password padlocks are blocked for non-whitelist anyway.