Allow room owners to block leashing (#1831)
* Allow room owners to block leashing

This adds a "Player Leashing" toggle to the room block menu. This does not block leashes by itself but simply prevents the "Hold On To Her Leash" option from appearing on players. This will not stop a dom from having a leashed sub and bringing her into the room and taking her out (unless one of them disconnects in the process), but it will stop people from grabbing the leash of another player without first asking the room owner to turn off the block.

* Updated ChatRoomCanBeLeashedBy