Feature: Ability to pick locks (#1675)
* Ability to pick locks

Oh boy, lots of changes in this one.
1) Instead of Inspect Lock and Unlock being in the menu at the same time, there is a "Lock Menu" which contains 3 options: Inspect, Unlock, and Pick
2) Lock picking minigame was added. it is very similar to the oblivion minigame for picking locks. Challenge increases based on lock time and it is possible to extend it to make it possible to pick locks while tied up, but currently you must be able to interact with items to do it.
3) Owner, Lover, Mistress locks are not pickable. The user can try, but there will be several pins that are impossible to set no matter what, even if the user gets extremely lucky.
Combination padlocks are pickable by those with lock picking skill of 5 or higher.
There is also a new High Security padlock. I plan to update this one to allow giving keys to others so they can unlock it as well. For now, it can only be opened by the person who locked it on, and can be picked with a lock picking skill of 4 or higher.

Currently it does not require lockpicks to pick a lock. I will be working on the feature soon, with key deposits taking away lockpicks as well, and owner settings taking away the right to own lockpicks. However, there will be a character action to temporarily lend a player your lockpicks until they leave the room, and to take away a bound player's lockpicks until they leave the room. Potentially this could backfire, if lending your lockpicks takes away your own ability to pick locks.

* Added high security padlock features

* Difficulty tweaks + remove lock jam

* Allow player to remove self as keyholder for high security lock

* Added ability to hand over keys

* Fixed lock property removal

* Intricate padlock is now exclusive, and must be picked.

* Bondage skill now matters, can attempt to pick lock with bound hands

If your hands are blocked, though, you cannot even attempt. This should prevent straitjackets from allowing people to pick locks while allowing someone with duct taped hands to feebly push a lock pick around with their hand. Mostly cosmetic as having bound hands adds +50 difficulty

* Added buttons to append whitelist, owner, lover to high security lock

* Difficulty tweak

* More difficulty tweaks + NotSelfPickable effect

* Player must now purchase lockpicks. Can be confiscated

* Fixed bugs with giving keys. Can also loan lockpicks now.

* Dialogue fixes

* Can now break loaned picks. Also, can pick with mittens if you have keys

* Fixes key requirement for mitten lock picking

* Difficulty tweaks

* Split character actions into Character and Security actions

* Fixed chastity belt crash + new setting to disable lockpicking on self

* Lockpicking is exclusive to the intricate lock and the high-sec lock

* Added arousal causing pins to reset, + fixed issue picking others

* Can now pick while blind + difficulty tweaks.

* Prevent lock menu showing up if cant pick lock

* Fixed some crashes

* Fixed crashes

* Added intricate and highsec locks to futuristic collar

* Fix high sec padlock inputs staying after updating them

* Fixed inputs to high sec padlock updating randomly..

* Intro text

* Fixed merge conflict