Screen flash on plug/vibe events (#1997)
* Messages when plugged or vibrated

Plays messages occasionally when you kneel/stand up, enter a chat room, use the struggle action, or fail a struggling minigame and exit out. 
These messages appear only to the player and not to others.

* Now adds arousal and hides if sex messages blocked

* Cleanup

* Made it not arouse QUITE as fast

* Immersion setting changes

* Fix for walking

* Screen flash on plug/vibe events

This is an addon to the messages while pluged or vibed PR, which makes the screen flash when the message happens. It can be turned off in the graphical menu.

* Switched to commontime to prevent glitches with flash lasting long