Strait Leotard Extended (#2587)
* Priority Update

A tiny change took actually a minute. I changed it so that the Strait Leotard has similar behaviour to the Futuristic Leotard.

That being that the suit itself is really low priority while the arms have a higher priority allowing items to be layered on top of the suit but below the arms.

* Update Female3DCG.js

Changed it from 23 to 31 so it goes over the majority of clothing items. Basically the jacket will remain under the outfit but the arms will go on top so it looks like you are hugging yourself over the clothes.

* Modular & Extended StraitLeotard

* Update BondageClub/Assets/Female3DCG/Female3DCGExtended.js

Co-authored-by: Ellie <62729616+elliesec@users.noreply.github.com>

* Update Female3DCG.js

* Update Female3DCGExtended.js

Co-authored-by: Ellie <62729616+elliesec@users.noreply.github.com>
Co-authored-by: Ben987 <30805159+Ben987@users.noreply.github.com>