Advanced Denial Modes (#2590)
* Enhanced Denial

* Fix

* Only happens while edged

* Advanced Denial

Adds a new feature to the Deny mode of most vibrators... ruined orgasms!

When the vibrator switches off, it will first edge the player and bring them to the orgasm screen. At this point there is nothing the player can do to orgasm, as the surrender button will be missing and the orgasm will cancel right before the timer runs out.

The mode is manually triggered by the vibrator. It will not be triggered in any other modes, so the change should be mostly safe.

This does not send a chat message to anyone but the player, in an effort to reduce chat spam in chat rooms. Deny mode will remain safe for use in rooms with 9 AFK dolls set to deny, for better or for worse.

Also adds it to the Sci Fi Pleasure Panties for more fun. I asked the creator for permission.

* Lint fix

* Additions

* Linting