Futuristic Item Update (#2606)
* Enhanced Denial

* Fix

* Only happens while edged

* Advanced Denial

Adds a new feature to the Deny mode of most vibrators... ruined orgasms!

When the vibrator switches off, it will first edge the player and bring them to the orgasm screen. At this point there is nothing the player can do to orgasm, as the surrender button will be missing and the orgasm will cancel right before the timer runs out.

The mode is manually triggered by the vibrator. It will not be triggered in any other modes, so the change should be mostly safe.

This does not send a chat message to anyone but the player, in an effort to reduce chat spam in chat rooms. Deny mode will remain safe for use in rooms with 9 AFK dolls set to deny, for better or for worse.

Also adds it to the Sci Fi Pleasure Panties for more fun. I asked the creator for permission.

* Lint fix

* Additions

* Linting

* Changes some futuristic stuff

Adds Futuristic Training belt, an exciting new item with a great deal of control over the wearer

Merges the futuristic chastity belts into one item to reduce clutter

Fixes an issue where the Futuristic Pleasure Panties showed both a remote icon and an extended item icon. Now it will only show the remote icon if the item is blocked.

Changed many futuristic items to use remotes instead of being AlwaysExtend, meaning that a player needs access to the controls if they are not allowed to use remotes.

Modified the futuristic vibrator to use the dark magic of regex to prevent single characters from tormenting people.

* Backup

* Edging and teasing

* Edge refinements

* Edgecontrol

* Imporovomentos

* Futuristic collar uses remote

* Add

* Speech punishments

* Final

Co-authored-by: Ben987 <30805159+Ben987@users.noreply.github.com>