Disable Examine While Blind only works outside of the virtual realm (#2575)
* Refactor custom draw list in ChatRoom.js

Refactors the custom drawing list in ChatRoom.js in order to support features besides the VR headset. Also fixes an issue with the orgasm filter.

* Disable Examine only works outside of the virtual realm

The virtual realm is supposed to allow players a bit more freedom with each other only, so it makes sense. Also, this allows switches to edit their sub's settings while in the virtual realm, even if they have immersion options set.